Rename `complete_daily_ecdr` module -> `intermediate_daily`
