+++ seaice_ecdr/cli/entrypoint.py
@@ -2,8 +2,6 @@
 
 import click
 
-from seaice_ecdr.complete_daily_ecdr import cli as complete_daily_cli
-
 # TODO: The daily-aggregate processing is very parallelizable because
 #       each year is indendent of every other year.  It could be
 #       implemented with multi-processing to speed up production
@@ -11,6 +9,7 @@ from seaice_ecdr.complete_daily_ecdr import cli as complete_daily_cli
 #       CLI API?
 from seaice_ecdr.daily_aggregate import cli as daily_aggregate_cli
 from seaice_ecdr.initial_daily_ecdr import cli as ecdr_cli
+from seaice_ecdr.intermediate_daily import cli as intermediate_daily_cli
 from seaice_ecdr.monthly import cli as monthly_cli
 from seaice_ecdr.monthly_aggregate import cli as monthly_aggregate_cli
 from seaice_ecdr.multiprocess_daily import cli as multiprocess_daily_cli
@@ -28,7 +27,7 @@ def cli():
 cli.add_command(ecdr_cli)
 cli.add_command(tiecdr_cli)
 cli.add_command(nrt_cli)
-cli.add_command(complete_daily_cli)
+cli.add_command(intermediate_daily_cli)
 cli.add_command(monthly_cli)
 cli.add_command(daily_aggregate_cli)
 cli.add_command(monthly_aggregate_cli)
