+++ seaice_ecdr/intermediate_daily.py
